backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
supabase==2.4.3
celery==5.4.0
redis==5.0.4
google-generativeai==0.5.4
pytesseract==0.3.10
pdf2image==1.17.0
PyMuPDF==1.24.3
Pillow==10.3.0
reportlab==4.1.0
httpx==0.27.0
resend==0.7.2
sentry-sdk[fastapi]==2.2.0
python-dotenv==1.0.1
aiofiles==23.2.1
cryptography==42.0.7
hashlib2==1.0.0
